Fixed some major issues with the complete hashing process.  Lots of assorted errors dealing with dict data being returned, missing modifiers, and some extra failsafe support for the spider.

+++ daemon.py
@@ -13,20 +13,21 @@ import hashtools
 from settings import MEDIA_ROOT
 
 def calculate_new_hashes():
-	for algo_name, func in hashtools.HASH_FUNCTIONS:
+	for algo_name, func in hashtools.HASH_ALGORITHMS.iteritems():
 		images_missing_hashes = database.get_images_without_hash(algo_name)
-		for img in images_missing_hashes:
-			logging.info("daemon.py: calculate_new_hashes: Processing algo {}, file {}".format(algo_name, img.filename))
-			img = Image.open(os.path.join(MEDIA_ROOT, img.filename))
-			hash = func(img)
-			database.create_hash(img.id, hash, algo_name)
+		for image_row in images_missing_hashes:
+			logging.info("daemon.py: calculate_new_hashes: Processing algo {}, file {}".format(algo_name, image_row['filename']))
+			image = Image.open(os.path.join(MEDIA_ROOT, image_row['filename']))
+			hash = func(image)
+			database.create_hash(image_row['id'], hash, algo_name)
 
 def clear_dead_links():
-	old_images = database.get_old_images(timeago=60*60*24*7, valid_entry=valid_entry)
+	old_images = database.get_old_images(timeago=60*60*24*7)
 	broken_image_ids = list()
 	for img in old_images:
 		pass # TODO: URL Check for 404
-	database.delete_images(broken_image_ids)
+	if len(broken_image_ids) > 0:
+		database.delete_images(broken_image_ids)
 
 if __name__=="__main__":
 	# TODO: Parallel launch
